:recycle: Refactor: add folder for not using scripts
